test: drop moto, flatten the e2e suite, and merge the four e2e workflows into one
The mocked control-plane suite is gone: floci drives the same three worker managers with real
workers booting, so moto was a second, weaker copy of the same coverage carrying its own AWS
compatibility shims. Removing it also removes the CI log's endless "--- Logging error ---" spam
(orb-py attaches a RotatingFileHandler to the ROOT logger under ORB_ROOT_DIR, which the test then
deleted with its temp dir; logging swallows handler errors, so it spammed stderr forever without
ever failing a test) and the Python 3.11 lane that existed only to run it.

The suite is now one flat, self-contained directory:

  * tests/integration/e2e/ folded up into tests/integration/; _harness/_floci/_container_*/
    _ecs_backend/_ec2_backend collapsed into harness/floci/docker/backends -- 18 modules to 9.
  * matrix.py's dynamic class generation replaced by one plain class per topology, so the whole
    matrix reads top to bottom.
  * one SCALER_E2E=<topology> gate replaces four RUN_*_E2E booleans, and probing is gone: asking
    for a topology whose prerequisites are missing now fails loudly instead of skipping green.
  * scripts moved to scripts/e2e/; four near-identical runners became run.sh <topology>.
  * four near-identical workflows became one label/dispatch-driven e2e.yml on Python 3.13.

Every scenario now prints its numbers on an [e2e] line before asserting, so a run's log says what
it proved; CI greps for them and fails a run that produced none.

Cut: work_spreads (its docstring conceded it proved only "more than one machine was used over the
run", which rising_load covers honestly), the single-implementation ContainerRuntime ABC, the
always-zero boot/shutdown delay simulation, Profile's 13 knobs down to 9, and the dead
ip_address/is_running/logs/object_storage_address/async_wait_until/desired_requests helpers.

Reverts tests/utility/utility.py: its additions were e2e-only and now live in the e2e harness.
