--- .github/workflows/cross-backend-e2e.yml
@@ -1,82 +1,0 @@
-name: "Floci Cross-Backend Waterfall E2E"
-
-# Full end-to-end cross-backend waterfall test (the Test_ecs_ec2 topology in tests/integration/e2e/): a
-# client bursts work at a real scheduler running the waterfall policy, driving the SHIPPED ECS worker
-# manager (priority 1) and the SHIPPED ORB/EC2 worker manager (priority 2) at once. Work fills the fast ECS
-# pool first and spills onto real Amazon Linux 2023 EC2 instances once ECS saturates, all via a local floci
-# AWS emulator. Each EC2 instance installs a manylinux wheel of the CURRENT source (built here with
-# cibuildwheel), so this is the heaviest job -- it builds that wheel and both task images and needs a Docker
-# daemon, so it is NOT part of the normal build-and-test run; trigger it on demand:
-#   * manually via the Actions tab ("Run workflow"); or
-#   * by adding the "cross-backend-e2e" label to a pull request.
-on:
-  workflow_dispatch:
-    inputs:
-      policy_content:
-        description: >-
-          Raw waterfall_v1 policy, one rule per line: 'priority,worker_manager_id[,max_task_concurrency]'.
-          Manager ids are wm-<name>-p<position>; for this topology: wm-ecs-p1, wm-ec2-p2. Blank = default
-          (ecs priority 1 capped at 2, ec2 priority 2).
-        required: false
-        type: string
-      num_tasks:
-        description: "Tasks per burst wave (SCALER_IT_NUM_TASKS). Blank = backend default."
-        required: false
-        type: string
-      task_seconds:
-        description: "Sleep per task, seconds (SCALER_IT_TASK_SECONDS). Blank = backend default (0.15)."
-        required: false
-        type: string
-  pull_request:
-    types: [ labeled ]
-
-permissions:
-  contents: read
-
-jobs:
-  cross-backend-e2e:
-    name: Floci Cross-Backend Waterfall E2E (Linux)
-    # Manual dispatch always runs; on a PR only when it carries the "cross-backend-e2e" label.
-    if: ${{ github.event_name == 'workflow_dispatch' || github.event.label.name == 'cross-backend-e2e' }}
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Harden the runner (Audit all outbound calls)
-        uses: step-security/harden-runner@8d3c67de8e2fe68ef647c8db1e6a09f647780f40 # v2.19.0
-        with:
-          egress-policy: audit
-
-      - name: 'Checkout Repository'
-        uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
-
-      - uses: ./.github/actions/setup-env
-        with:
-          os: ${{ runner.os }}
-          python-version: '3.13'
-
-      - uses: ./.github/actions/build-native-dependencies
-        with:
-          os: ${{ runner.os }}
-
-      # Install scaler on the host so both manager processes (the shipped ECS + ORB/EC2 managers) run.
-      - uses: ./.github/actions/build-and-install-wheel
-
-      # The EC2 instances run the CURRENT source, so build a portable manylinux wheel of it (not a PyPI
-      # release). Matches the interpreter the instances install (cp313).
-      - name: Build current-source manylinux wheel
-        shell: bash
-        run: |
-          CIBW_OUTPUT_DIR=dist_manylinux ./scripts/build_cibuildwheel.sh --only cp313-manylinux_x86_64
-
-      - name: Run floci cross-backend waterfall e2e
-        shell: bash
-        env:
-          RUN_CROSS_BACKEND_E2E: '1'
-          # GitHub runners can reach the Docker socket without sudo.
-          SCALER_IT_CONTAINER_CLI: docker
-          # workflow_dispatch knobs; empty on the labeled-PR trigger, which the framework reads as "unset".
-          SCALER_IT_WATERFALL_POLICY: ${{ github.event.inputs.policy_content }}
-          SCALER_IT_NUM_TASKS: ${{ github.event.inputs.num_tasks }}
-          SCALER_IT_TASK_SECONDS: ${{ github.event.inputs.task_seconds }}
-        run: |
-          python -m unittest tests.integration.e2e.test_scaling -v
