--- .github/workflows/floci-e2e.yml
@@ -1,64 +1,0 @@
-name: "Floci ECS Scaling E2E"
-
-# Full end-to-end ECS-scaling test (the Test_ecs topology in tests/integration/e2e/): a client bursts work at a
-# real scheduler, whose scaling policy drives the SHIPPED ECS worker manager to launch real workers in real
-# Docker containers via a local floci AWS emulator (a free stand-in for AWS that actually runs RunTask
-# containers). It builds an ECS task image and needs a Docker daemon, so it is NOT part of the normal
-# build-and-test run; trigger it on demand:
-#   * manually via the Actions tab ("Run workflow"); or
-#   * by adding the "floci-e2e" label to a pull request.
-on:
-  workflow_dispatch:
-    inputs:
-      num_tasks:
-        description: "Tasks per burst wave (SCALER_IT_NUM_TASKS). Blank = backend default."
-        required: false
-        type: string
-      task_seconds:
-        description: "Sleep per task, seconds (SCALER_IT_TASK_SECONDS). Blank = backend default (0.15)."
-        required: false
-        type: string
-  pull_request:
-    types: [ labeled ]
-
-permissions:
-  contents: read
-
-jobs:
-  floci-e2e:
-    name: Floci ECS Scaling E2E (Linux)
-    # Manual dispatch always runs; on a PR only when it carries the "floci-e2e" label.
-    if: ${{ github.event_name == 'workflow_dispatch' || github.event.label.name == 'floci-e2e' }}
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Harden the runner (Audit all outbound calls)
-        uses: step-security/harden-runner@8d3c67de8e2fe68ef647c8db1e6a09f647780f40 # v2.19.0
-        with:
-          egress-policy: audit
-
-      - name: 'Checkout Repository'
-        uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
-
-      - uses: ./.github/actions/setup-env
-        with:
-          os: ${{ runner.os }}
-          python-version: '3.11'
-
-      - uses: ./.github/actions/build-native-dependencies
-        with:
-          os: ${{ runner.os }}
-
-      - uses: ./.github/actions/build-and-install-wheel
-
-      - name: Run floci ECS scaling e2e
-        shell: bash
-        env:
-          RUN_FLOCI_E2E: '1'
-          # GitHub runners can reach the Docker socket without sudo.
-          SCALER_IT_CONTAINER_CLI: docker
-          # workflow_dispatch knobs; empty on the labeled-PR trigger, which the framework reads as "unset".
-          SCALER_IT_NUM_TASKS: ${{ github.event.inputs.num_tasks }}
-          SCALER_IT_TASK_SECONDS: ${{ github.event.inputs.task_seconds }}
-        run: |
-          python -m unittest tests.integration.e2e.test_scaling -v
