--- tests/integration/_harness.py
@@ -1,334 +1,0 @@
-"""Shared harness for the container-scaling e2e and the AWS control-plane integration tests.
-
-Brings up a real object-storage server + scheduler (in their own processes, exactly like production)
-with clean, exception-safe teardown, plus the small polling helper and the scheduler-command helper.
-Optionally starts the web GUI wired to the scheduler's monitor address so a local run can be watched in
-a browser. Kept separate from the AWS backend so the same harness serves both the AWS control-plane
-tests and the container-scaling data-plane tests.
-"""
-
-from __future__ import annotations
-
-import asyncio
-import os
-import signal
-import socket
-import tempfile
-import time
-from multiprocessing import get_context
-from multiprocessing.process import BaseProcess
-from typing import Callable, Optional
-
-from scaler.cluster.object_storage_server import ObjectStorageServerProcess
-from scaler.cluster.scheduler import SchedulerProcess
-from scaler.config.defaults import (
-    DEFAULT_CLIENT_TIMEOUT_SECONDS,
-    DEFAULT_IO_THREADS,
-    DEFAULT_LOAD_BALANCE_SECONDS,
-    DEFAULT_LOAD_BALANCE_TRIGGER_TIMES,
-    DEFAULT_MAX_NUMBER_OF_TASKS_WAITING,
-    DEFAULT_OBJECT_RETENTION_SECONDS,
-    DEFAULT_WORKER_TIMEOUT_SECONDS,
-)
-from scaler.config.section.scheduler import PolicyConfig
-from scaler.config.types.address import AddressConfig
-from scaler.scheduler.controllers.worker_manager_utilties import build_set_desired_command
-from scaler.utility.network_util import get_available_tcp_port
-from tests.utility.utility import POLL_INTERVAL_SECONDS, PROCESS_TERMINATION_TIMEOUT_SECONDS
-
-DEFAULT_WAIT_TIMEOUT_SECONDS = 30.0
-# get_available_tcp_port() releases the port before the child binds, so a since-freed ephemeral port
-# can be grabbed in between (a TOCTOU race). Retry the whole bring-up with fresh ports a few times.
-_HARNESS_START_ATTEMPTS = 3
-
-
-async def async_wait_until(
-    predicate: Callable[[], bool], timeout: float = DEFAULT_WAIT_TIMEOUT_SECONDS, message: str = ""
-) -> None:
-    """Poll ``predicate``, yielding to the event loop between checks, until it is true or ``timeout``
-    elapses (then raise ``TimeoutError``)."""
-    deadline = time.monotonic() + timeout
-    while time.monotonic() < deadline:
-        if predicate():
-            return
-        await asyncio.sleep(POLL_INTERVAL_SECONDS)
-    raise TimeoutError(message or f"condition not met within {timeout:.1f}s")
-
-
-def desired_requests(count: int) -> list:
-    """The exact per-capset request list the scheduler's scaling policy sends over the wire for a target
-    concurrency (shared by the ECS and EC2/ORB control-plane tests)."""
-    return list(build_set_desired_command([({}, count)]).setDesiredTaskConcurrencyRequests)
-
-
-def assert_backend_processes_alive(test_case, harness: "SchedulerHarness", **named_processes) -> None:
-    """Fail ``test_case`` with a clear diagnostic if the scheduler or any named worker-manager process died,
-    or if the scheduler logged an unhandled exception, during the test. A client that fails with a bare
-    ``TimeoutError`` is almost always the *downstream* symptom of one of those under churn -- whether the
-    scheduler crashed outright or stayed up but wedged the pool -- so naming the fault (and pointing at the
-    traceback the harness captures) turns it into an actionable failure. Call from tearDown, which runs
-    before the addCleanup stack tears the processes down."""
-    dead = []
-    if harness.scheduler_died():
-        dead.append(f"scheduler (exit {harness.scheduler_exitcode()})")
-    if harness.object_storage_died():
-        dead.append(f"object storage (exit {harness.object_storage_exitcode()})")
-    for name, process in named_processes.items():
-        if process is not None and not process.is_alive():
-            dead.append(f"{name} (exit {process.exitcode})")
-    # The scheduler can log its crash traceback a beat after the client gives up, so wait for it to land only
-    # when a fault is already evident (a process died); a green run scans the log once and returns at once.
-    scheduler_error = harness.scheduler_unhandled_error(settle_seconds=4.0 if dead else 0.0)
-    if dead:
-        detail = f" scheduler logged: {scheduler_error}." if scheduler_error else ""
-        test_case.fail(
-            f"{', '.join(dead)} exited during the run -- an unhandled crash under churn, not just a slow or "
-            f"timed-out client. See the traceback in the captured scheduler/manager log above for the cause." + detail
-        )
-    elif scheduler_error:
-        test_case.fail(
-            f"scheduler logged an unhandled exception under churn and wedged the pool ({scheduler_error}); "
-            f"the client's TimeoutError is the downstream symptom. See the full traceback in the captured log."
-        )
-
-
-def _run_webgui(monitor_address: str, host: str, port: int) -> None:
-    """Process entry point for the web GUI (blocking uvicorn server); it subscribes to the scheduler's
-    monitor address and serves the dashboard on ``host:port``."""
-    from scaler.config.section.webgui import WebGUIConfig
-    from scaler.config.types.address import AddressConfig as _AddressConfig
-    from scaler.config.types.http import HTTPConfig
-    from scaler.ui.webgui import start_webgui
-
-    start_webgui(
-        WebGUIConfig(monitor_address=_AddressConfig.from_string(monitor_address), gui_address=HTTPConfig(host, port))
-    )
-
-
-class SchedulerHarness:
-    """A running object-storage server + scheduler, addressed over loopback TCP.
-
-    Unlike ``SchedulerClusterCombo`` this does NOT start any workers of its own, so a test can attach
-    its own (dynamic / cloud) worker manager and observe the full provisioning control loop. Register
-    :meth:`shutdown` with ``self.addCleanup`` so it is torn down even if the test body raises.
-    """
-
-    def __init__(
-        self,
-        policy_content: str = "allocate=even_load; scaling=vanilla",
-        policy_engine_type: str = "simple",
-        gateway: Optional[str] = None,
-        enable_webgui: bool = False,
-        client_timeout_seconds: int = DEFAULT_CLIENT_TIMEOUT_SECONDS,
-        worker_timeout_seconds: int = DEFAULT_WORKER_TIMEOUT_SECONDS,
-    ) -> None:
-        # gateway: a docker-bridge gateway IP. When set, the scheduler + object storage bind 0.0.0.0 and
-        # advertise gateway-reachable addresses so workers running in containers can connect back; the
-        # host client still uses loopback. None keeps the original loopback-only behavior.
-        self._gateway = gateway
-        self._enable_webgui = enable_webgui
-        # Cloud provisioning latency (a real EC2 boot is minutes) can starve the scheduler and stall client
-        # heartbeats past the stock 60s, so the floci e2es raise these; loopback tests keep the defaults.
-        self._client_timeout_seconds = client_timeout_seconds
-        self._worker_timeout_seconds = worker_timeout_seconds
-        self._object_storage: Optional[ObjectStorageServerProcess] = None
-        self._scheduler: Optional[SchedulerProcess] = None
-        self._webgui: Optional[BaseProcess] = None
-        self._monitor_client_address: Optional[str] = None
-        self._shutdown_done = False
-        # Tee the scheduler log to a file (as well as stdout) so a test can detect an unhandled scheduler
-        # exception under churn -- which may crash the scheduler OR leave it up but wedged (client times out).
-        scheduler_log_fd, self._scheduler_log_path = tempfile.mkstemp(prefix="scaler-it-scheduler-", suffix=".log")
-        os.close(scheduler_log_fd)
-
-        errors: list = []
-        for _attempt in range(_HARNESS_START_ATTEMPTS):
-            try:
-                self._bring_up(policy_content, policy_engine_type)
-                break
-            except (OSError, TimeoutError) as error:
-                # A since-freed ephemeral port was grabbed before a child could bind it; tear the
-                # partial bring-up down and retry with fresh ports.
-                errors.append(error)
-                self._kill_processes()
-        else:
-            raise RuntimeError(
-                f"SchedulerHarness failed to start after {_HARNESS_START_ATTEMPTS} attempts: {errors!r}"
-            ) from errors[-1]
-
-        if self._enable_webgui:
-            self._start_webgui()
-
-    def _bring_up(self, policy_content: str, policy_engine_type: str) -> None:
-        sched_port, os_port = get_available_tcp_port(), get_available_tcp_port()
-        bind_host = "0.0.0.0" if self._gateway else "127.0.0.1"
-        # The host client always connects over loopback; container workers use the gateway address.
-        self.scheduler_address = f"tcp://127.0.0.1:{sched_port}"
-        self.worker_scheduler_address = (
-            f"tcp://{self._gateway}:{sched_port}" if self._gateway else self.scheduler_address
-        )
-        self._object_storage_address = AddressConfig.from_string(f"tcp://{bind_host}:{os_port}")
-        advertised_os = AddressConfig.from_string(f"tcp://{self._gateway}:{os_port}") if self._gateway else None
-
-        monitor_address = None
-        if self._enable_webgui:
-            monitor_port = get_available_tcp_port()
-            monitor_address = AddressConfig.from_string(f"tcp://{bind_host}:{monitor_port}")
-            self._monitor_client_address = f"tcp://127.0.0.1:{monitor_port}"
-
-        self._object_storage = ObjectStorageServerProcess(
-            bind_address=self._object_storage_address,
-            identity="ObjectStorageServer",
-            logging_paths=("/dev/stdout",),
-            logging_config_file=None,
-            logging_level="INFO",
-        )
-        self._object_storage.start()
-        self._object_storage.wait_until_ready()  # raises TimeoutError if the port was taken
-
-        self._scheduler = SchedulerProcess(
-            bind_address=AddressConfig.from_string(f"tcp://{bind_host}:{sched_port}"),
-            object_storage_address=AddressConfig.from_string(f"tcp://127.0.0.1:{os_port}"),
-            advertised_object_storage_address=advertised_os,
-            monitor_address=monitor_address,
-            policy=PolicyConfig(policy_engine_type=policy_engine_type, policy_content=policy_content),
-            io_threads=DEFAULT_IO_THREADS,
-            max_number_of_tasks_waiting=DEFAULT_MAX_NUMBER_OF_TASKS_WAITING,
-            client_timeout_seconds=self._client_timeout_seconds,
-            worker_timeout_seconds=self._worker_timeout_seconds,
-            object_retention_seconds=DEFAULT_OBJECT_RETENTION_SECONDS,
-            load_balance_seconds=DEFAULT_LOAD_BALANCE_SECONDS,
-            load_balance_trigger_times=DEFAULT_LOAD_BALANCE_TRIGGER_TIMES,
-            protected=False,
-            event_loop="builtin",
-            logging_paths=("/dev/stdout", self._scheduler_log_path),
-            logging_config_file=None,
-            logging_level="INFO",
-        )
-        self._scheduler.start()
-        self._wait_scheduler_ready(sched_port)
-
-    def _wait_scheduler_ready(self, port: int, timeout: float = DEFAULT_WAIT_TIMEOUT_SECONDS) -> None:
-        """Fail bring-up if the scheduler did not come up on ``port``.
-
-        get_available_tcp_port() releases the port before the child binds it, so a foreign process can grab
-        it in between (the TOCTOU the retry loop exists for); the scheduler then dies on its bind. Only the
-        object storage port is otherwise probed (wait_until_ready), so without this the scheduler race slips
-        through as a dead scheduler that surfaces much later as a client TimeoutError misattributed to a
-        crash under churn. Raising TimeoutError here routes it back through the fresh-ports retry instead."""
-        deadline = time.monotonic() + timeout
-        while time.monotonic() < deadline:
-            if self._scheduler is not None and not self._scheduler.is_alive():
-                raise TimeoutError(f"scheduler exited during startup (exit {self._scheduler.exitcode}); port taken?")
-            try:
-                with socket.create_connection(("127.0.0.1", port), timeout=1):
-                    return  # scheduler process is alive and its port accepts connections
-            except OSError:
-                time.sleep(POLL_INTERVAL_SECONDS)
-        raise TimeoutError(f"scheduler did not accept connections on 127.0.0.1:{port} within {timeout:.0f}s")
-
-    def _start_webgui(self) -> None:
-        """Start the dashboard, wired to the scheduler monitor, and print its URL. Best-effort: neither a
-        missing ``[gui]`` extra nor a failed spawn/port grab may fail the e2e (the GUI is a convenience, not
-        the thing under test) -- and it runs after the scheduler is up, so a raise here would leak it."""
-        try:
-            import scaler.ui.webgui  # noqa: F401
-
-            port = get_available_tcp_port()
-            webgui = get_context("spawn").Process(
-                target=_run_webgui, args=(self._monitor_client_address, "0.0.0.0", port), daemon=True
-            )
-            webgui.start()
-        except Exception as error:
-            print(f"[harness] web GUI unavailable ({error}); continuing without it")
-            return
-        self._webgui = webgui
-        print(f"[harness] web GUI: http://localhost:{port}  (scheduler monitor {self._monitor_client_address})")
-
-    @property
-    def object_storage_address(self) -> AddressConfig:
-        return self._object_storage_address
-
-    def scheduler_died(self) -> bool:
-        """True if the scheduler process has exited. A healthy run keeps it alive for the whole test, so a
-        dead scheduler mid-test means an unhandled scheduler crash (see :func:`assert_backend_processes_alive`)."""
-        return self._scheduler is not None and not self._scheduler.is_alive()
-
-    def scheduler_exitcode(self) -> Optional[int]:
-        return self._scheduler.exitcode if self._scheduler is not None else None
-
-    def object_storage_died(self) -> bool:
-        """True if the object-storage server exited mid-test (it logs only to stdout, so its crash is
-        otherwise invisible to the scheduler-log scan)."""
-        return self._object_storage is not None and not self._object_storage.is_alive()
-
-    def object_storage_exitcode(self) -> Optional[int]:
-        return self._object_storage.exitcode if self._object_storage is not None else None
-
-    def scheduler_unhandled_error(self, settle_seconds: float = 0.0) -> Optional[str]:
-        """A one-line summary if the scheduler logged an unhandled exception (a traceback) during the run,
-        else None -- this catches a fault that wedges the pool without killing the process. ``settle_seconds``
-        > 0 briefly polls for a traceback that lands just after the scan (the caller sets it only when a crash
-        is already evident). A silent wedge that logs NO traceback (a deadlock, or a C++/ymq-layer fault) is
-        out of reach here; only ``scheduler_died`` and the scenario's own timeout guard that. Full traceback
-        is in the captured log."""
-        deadline = time.monotonic() + settle_seconds
-        while True:
-            try:
-                with open(self._scheduler_log_path, "r", encoding="utf-8", errors="replace") as log_file:
-                    lines = log_file.read().splitlines()
-            except OSError:
-                return None
-            if any("Traceback (most recent call last)" in line for line in lines):
-                # __routing logs "<task>: exception happened, transition: ... path: ..." right before the
-                # traceback; report that line (it names the state path) if present, else the exception line.
-                context = next((line.strip() for line in lines if "exception happened" in line), "")
-                exception = next(
-                    (line.strip() for line in reversed(lines) if line[:1].strip() and "Error" in line and ":" in line),
-                    "",
-                )
-                return " | ".join(part for part in (context, exception) if part) or "unhandled scheduler exception"
-            if time.monotonic() >= deadline:
-                return None
-            time.sleep(0.25)
-
-    def _kill_processes(self) -> None:
-        """Force-terminate whatever is running (used to clean up a partial bring-up between retries)."""
-        for process in (self._scheduler, self._object_storage):
-            if process is not None and process.is_alive():
-                process.terminate()
-                process.join(timeout=PROCESS_TERMINATION_TIMEOUT_SECONDS)
-                if process.is_alive():
-                    process.kill()
-                    process.join()
-        self._scheduler = None
-        self._object_storage = None
-
-    def shutdown(self) -> None:
-        if self._shutdown_done:
-            return
-        self._shutdown_done = True
-        if self._webgui is not None and self._webgui.is_alive():
-            self._webgui.terminate()
-            self._webgui.join(timeout=PROCESS_TERMINATION_TIMEOUT_SECONDS)
-            if self._webgui.is_alive():
-                self._webgui.kill()
-                self._webgui.join()
-        for process, use_sigint in ((self._scheduler, True), (self._object_storage, False)):
-            if process is None or not process.is_alive():
-                continue
-            try:
-                if use_sigint:
-                    os.kill(process.pid, signal.SIGINT)
-                else:
-                    process.terminate()
-            except (ProcessLookupError, OSError):
-                pass
-            process.join(timeout=PROCESS_TERMINATION_TIMEOUT_SECONDS)
-            if process.is_alive():
-                process.kill()
-                process.join()
-        try:
-            os.unlink(self._scheduler_log_path)
-        except OSError:
-            pass
