+++ tests/integration/worker.Dockerfile
@@ -1,10 +1,10 @@
-# Self-contained worker image for the container-scaling e2e (see tests/integration/README.md).
+# Self-contained worker image for the e2e scaling tests (see README.md).
 #
 # It carries the exact wheel the host just built plus the custom-built capnp/kj shared libs the
 # scaler C++ extensions link against, so a container runs the same scaler build and wire protocol
 # as the host scheduler -- with none of the host-layout coupling a bind-mount would need.
 #
-# Built by tests/integration/_container_image.py, which stages the wheel under wheel/ and the libs
+# Built by docker.py, which stages the wheel under wheel/ and the libs
 # under libs/. BASE_IMAGE must have a glibc at least as new as the host that built the wheel/libs
 # (hence ubuntu:26.04 by default, to match the dev host); PYTHON_VERSION must match the wheel ABI.
 ARG BASE_IMAGE=ubuntu:26.04
